move hwi_server.py
