+++ src/cryptoadvance/specter/server.py
@@ -28,7 +28,7 @@ from cryptoadvance.specter.services import callbacks
 from cryptoadvance.specter.util.reflection import get_template_static_folder
 
 from .htmlsafebabel import HTMLSafeBabel
-from .hwi_server import hwi_server
+from ..specterext.hwi.hwi_server import hwi_server
 from .services.callbacks import after_serverpy_init_app, specter_added_to_flask_app
 from .specter import Specter
 from .util.specter_migrator import SpecterMigrator
