+++ src/cryptoadvance/specterext/hwi/helpers.py
@@ -1,0 +1,29 @@
+import os
+from cryptoadvance.specter.persistence import read_json_file, write_json_file
+from cryptoadvance.specter.helpers import deep_update
+
+
+def hwi_get_config(specter):
+    config = {"whitelisted_domains": "http://127.0.0.1:25441/"}
+    # if hwi_bridge_config.json file exists - load from it
+    if os.path.isfile(os.path.join(specter.data_folder, "hwi_bridge_config.json")):
+        fname = os.path.join(specter.data_folder, "hwi_bridge_config.json")
+        file_config = read_json_file(fname)
+        deep_update(config, file_config)
+    # otherwise - create one and assign unique id
+    else:
+        save_hwi_bridge_config(specter, config)
+    return config
+
+
+def save_hwi_bridge_config(specter, config):
+    if "whitelisted_domains" in config:
+        whitelisted_domains = ""
+        for url in config["whitelisted_domains"].split():
+            if not url.endswith("/") and url != "*":
+                # make sure the url end with a "/"
+                url += "/"
+            whitelisted_domains += url.strip() + "\n"
+        config["whitelisted_domains"] = whitelisted_domains
+    fname = os.path.join(specter.data_folder, "hwi_bridge_config.json")
+    write_json_file(config, fname)
